moved secrets to secrets.json
