+++ .gitignore
@@ -1,2 +1,2 @@
-src/token.json
+src/secrets.json
 node_modules
