+++ README.md
@@ -10,6 +10,13 @@ A bot to monitor LiveOverflow's minecraft server using discord.
 Create src/token.json with contents
 ```json
 {
-    "token": "your-discord-bot-token"
+    "token": "...",
+    "botid": "...",
+    "connectionData": {
+        "host": "...",
+        "port": 25565,
+        "username": "Player",
+        "password": "password"
+    }
 }
 ```
