+++ src/minecraft.json
@@ -19,11 +19,5 @@
         "pink_stained_glass",
         "light_blue_stained_glass",
         "white_stained_glass"
-    ],
-    "connectionData": {
-        "host": "localhost",
-        "port": 25565,
-        "username": "Player",
-        "password": "password"
-    }
+    ]
 }
